lib_java_dla-tr7493-1

+++ java/org/jax/mgi/dbs/mgd/loads/Coord/CoordinateInputProcessor.java
@@ -3,6 +3,7 @@ package org.jax.mgi.dbs.mgd.loads.Coord;
 import org.jax.mgi.shr.config.CoordLoadCfg;
 import org.jax.mgi.shr.config.ConfigException;
 import org.jax.mgi.shr.dbutils.DBException;
+import org.jax.mgi.dbs.mgd.lookup.TranslationException;
 import org.jax.mgi.shr.cache.CacheException;
 import org.jax.mgi.dbs.mgd.lookup.TranslationException;
 import org.jax.mgi.shr.cache.KeyNotFoundException;
@@ -12,6 +13,7 @@ import org.jax.mgi.dbs.mgd.dao.MAP_Coord_CollectionState;
 import org.jax.mgi.dbs.mgd.dao.MAP_Coord_CollectionDAO;
 import org.jax.mgi.dbs.mgd.loads.Coord.CoordMapProcessor;
 import org.jax.mgi.dbs.mgd.loads.Coord.CoordMapFeatureResolver;
+import org.jax.mgi.dbs.mgd.lookup.CoordMapCollectionKeyLookup;
 import org.jax.mgi.shr.dla.loader.coord.CoordloaderExceptionFactory;
 import org.jax.mgi.shr.dla.input.CoordinateInput;
 import org.jax.mgi.shr.dla.log.DLALogger;
@@ -20,6 +22,8 @@ import org.jax.mgi.shr.dla.loader.coord.CoordloaderException;
 import org.jax.mgi.shr.dbutils.dao.SQLStream;
 import org.jax.mgi.shr.dbutils.SQLDataManager;
 import org.jax.mgi.shr.dbutils.SQLDataManagerFactory;
+import org.jax.mgi.shr.dbutils.ResultsNavigator;
+import org.jax.mgi.shr.dbutils.RowReference;
 import org.jax.mgi.dbs.SchemaConstants;
 
 /**
@@ -47,7 +51,7 @@ import org.jax.mgi.dbs.SchemaConstants;
 
 public class CoordinateInputProcessor {
     // gets configuration values for coordinate loads
-    private CoordLoadCfg coordCfg;
+    private CoordLoadCfg loadCfg;
 
     // name of the collection
     private String collectionName;
@@ -55,14 +59,37 @@ public class CoordinateInputProcessor {
     // abbreviation for the collection
     private String collectionAbbrev;
 
+    // delete_reload or incremental
+    private String loadMode;
+ 
+    // count of current number of features found to already be in MGI
+    private int featureInMGICtr = 0;
+
+    // count of current number of features found to be in MGI, but with
+    // different attribute(s)
+    private int featureDiffInMGICtr = 0;
+
+    // count of current number of feature objects (e.g. sequence/marker)
+    // not found in MGI
+    private int featureObjectNotInMGICtr = 0;
+
+    // true if collection is not in the database
+    private Boolean isNewCollection;
+
+    // Lookup a collection key
+    private CoordMapCollectionKeyLookup collectionLookup;
+
     // the collection key for this load
     private Integer collectionKey;
 
     // gets existing or creates the coordinate map for the coordinate
     private CoordMapProcessor mapProcessor;
 
-    // resolves CoordMapFeatureRqwAttributes to a MAP_Coord_FeatureState
+    // resolves CoordMapFeatureRawAttributes to a MAP_Coord_FeatureState
     private CoordMapFeatureResolver featureResolver;
+    
+    // needed to execute Queries
+    private SQLDataManager sqlMgr;
 
     // a stream for handling MGD DAO objects
     private SQLStream mgdStream;
@@ -70,52 +97,53 @@ public class CoordinateInputProcessor {
     // a coordinate Exception Factory
     private CoordloaderExceptionFactory eFactory;
     
-    DLALogger logger;
+    private DLALogger logger;
     /**
      * Constructs a CoordinateInputProcessor
      * @param stream stream for adding coordinates to an  MGD databaseth
      * @throws CacheException if error creating CoordMapFeatureResolver
      * @throws DBException if error creating CoordMapFeatureResolver
      * @throws ConfigException if error creating configurator or collection name
-     *          not found. Note  that coordCfg.getMapCollectionAbbrev supplies
+     *          not found. Note  that loadCfg.getMapCollectionAbbrev supplies
      *          a default value if not configured.
      * @throws KeyNotFoundException if error creating CoordMapFeatureResolver
      */
 
     public CoordinateInputProcessor(SQLStream stream) throws DBException, CacheException,
-        ConfigException, KeyNotFoundException, DLALoggingException {
-
+        ConfigException, KeyNotFoundException, DLALoggingException, TranslationException,
+	CoordloaderException {
+   
         mgdStream = stream;
+	
+	sqlMgr = SQLDataManagerFactory.getShared(SchemaConstants.MGD);
         eFactory = new CoordloaderExceptionFactory();
-        coordCfg = new CoordLoadCfg();
-        collectionName = coordCfg.getMapCollectionName();
-        collectionAbbrev = coordCfg.getMapCollectionAbbrev();
+	
+	// set values from configuration
+        loadCfg = new CoordLoadCfg();
+        collectionName = loadCfg.getMapCollectionName();
+	loadMode = loadCfg.getLoadMode();
+	collectionAbbrev = loadCfg.getMapCollectionAbbrev();
 
-        // set collection abbreviation to the name value if not configured
+        // set collection abbreviation to the collectionName value if not configured
         if(collectionAbbrev == null || collectionAbbrev.equals("")) {
                 collectionAbbrev = collectionName;
         }
         // create an instance of a CoordMapProcessor from configuration
-        mapProcessor = (CoordMapProcessor)coordCfg.getMapProcessorClass();
-
-        featureResolver = new CoordMapFeatureResolver();
+        mapProcessor = (CoordMapProcessor)loadCfg.getMapProcessorClass();
+	
+	// delete reload load mode
+	if (loadMode.equals("delete_reload")) {
+	    // this removes the collection too
+	    deleteCoordinates();
+	}
+	// find the collection in mgi or create one
+	determineCollection();
+	
+	featureResolver = new CoordMapFeatureResolver(isNewCollection);
 	logger = DLALogger.getInstance();
     }
-    /**
-     * deletes the collection, all coordinate maps and features for the
-     * collection. Creates a new collection object and sets the collection key
-     * in the map processor
-     * @throws CoordloaderException if error deleting collectin and associated
-     *    maps and features
-     * @throws ConfigException if error creating new collection
-     * @throws DBException if error creating new collection
-     */
-    public void preprocess() throws CoordloaderException, ConfigException,
-        DBException {
-        deleteCoordinates();
-        createCollection();
-    }
-
+    
+    
      /**
      * Adds a Coordinate Collection, Coordinate Maps for the Collections and
      * Coordinate Features to the database
@@ -141,22 +169,59 @@ public class CoordinateInputProcessor {
                 input.getCoordMapRawAttributes(), coordinate);
 
         // resolve the feature
-	MAP_Coord_FeatureState state; 
+	MAP_Coord_FeatureState state = null; 
 	try {
             state = featureResolver.resolve(
             input.getCoordMapFeatureRawAttributes(), mapKey);
 	}
-	catch (KeyNotFoundException e) {
-	    logger.logcInfo(e.getMessage(), true);
+	catch (FeatureInMGIException e1) {
+	    // There is already a Feature for this object ID in MGI
+	    featureInMGICtr++;
+	    return;
+	}
+	catch (FeatureDifferentInMGIException e2) {
+            // There is a Feature in MGI with different attributes 
+	    // for this object ID 
+            featureDiffInMGICtr++;
+            return;
+        }
+	catch (FeatureObjectNotInMGIException e3) {
+	    // This object ID does not exist in the database
+	    featureObjectNotInMGICtr++;
+	    logger.logvInfo(e3.getMessage(), true);
     	    return;
 	}
 
-        // set the feature in the coordMap object
+        // set the feature in the coordinate object
         coordinate.setCoordMapFeatureState(state);
 
         // send the CoordinateMap object to its stream
         coordinate.sendToStream();
     }
+    /**
+     * get the number of features already in MGI
+     *
+     */
+	public int getFeatureInMGICount() {
+	    return featureInMGICtr;
+	}
+    /**
+     * get the number of features already in MGI, but with different attributes
+     *
+     */
+        public int getFeatureDiffInMGICount() {
+            return featureDiffInMGICtr;
+        }
+
+
+    /**
+     * get the number of feature objects not in MGI
+     *
+     */
+        public int getFeatureObjectNotInMGICount() {
+            return featureObjectNotInMGICtr;
+        }
+
 
     /**
      * deletes the coordinate collection, all coordinate maps and features for the
@@ -171,7 +236,6 @@ public class CoordinateInputProcessor {
 
        String spCall = "MAP_deleteByCollection '" + collectionName + "'";
        try {
-         SQLDataManager sqlMgr = SQLDataManagerFactory.getShared(SchemaConstants.MGD);
          sqlMgr.executeSimpleProc(spCall);
        }
        catch (MGIException e) {
@@ -182,12 +246,27 @@ public class CoordinateInputProcessor {
        }
      }
 
+    private void determineCollection () throws  DBException, ConfigException, 
+	    CacheException {
+	    CoordMapCollectionKeyLookup collectionLookup = 
+		new CoordMapCollectionKeyLookup();
+	    collectionKey = collectionLookup.lookup(collectionName);
+	    
+	    if (collectionKey == null) {
+		mapProcessor.setCollectionKey(createCollection());
+		isNewCollection = Boolean.TRUE;
+	    }
+	    else {
+		mapProcessor.setCollectionKey(collectionKey);
+		isNewCollection = Boolean.FALSE;
+	    }
+    }     
     /**
      * Creates a collection object and sets the collection key in the map processor
      * @throws DBException if errors creating or inserting collection DAO
      * @throws ConfigException if error collection DAO
      */
-    private void createCollection () throws  DBException, ConfigException {
+    private Integer createCollection () throws  DBException, ConfigException {
             // create the state
             MAP_Coord_CollectionState collection = new MAP_Coord_CollectionState();
 
@@ -203,8 +282,7 @@ public class CoordinateInputProcessor {
 
             // insert the collection
             mgdStream.insert(dao);
-        // we have to explicitly set the collection key in the CoordMapProcessor
-        // because it is configured; Configurator does not take parameters
-        mapProcessor.setCollectionKey(collectionKey);
+        
+	    return collectionKey;
     }
 }
